Load_Testing_Streamlit_Apps/requirements.txt
streamlit==0.88.0
pandas==1.2.4
numpy==1.19.5
locustio==0.999
